pyproject.toml
[project]
name = "tg_analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
package-mode = false
dependencies = [
    "aerich>=0.9.0",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "gunicorn>=23.0.0",
    "httptools>=0.6.4",
    "passlib>=1.7.4",
    "pydantic>=2.11.3",
    "pydantic-settings>=2.8.1",
    "pyjwt>=1.7.1",
    "pysocks>=1.7.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "structlog>=25.2.0",
    "toml>=0.10.2",
    "tortoise-orm>=0.25.0",
    "uvicorn>=0.34.2",
    "uvloop>=0.21.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.15.0",
    "ruff>=0.11.5",
]
